paas-worker/bocloud_worker/requirements.txt
Flask
flask_restful
pika==0.10.0
kazoo
pywinrm
pyghmi
pysnmp
requests
passlib>=1.7.1
pycrypto
